made it serverless for vercel
